requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-imagekit==4.1.0
Pillow==9.4.0
psycopg2==2.9.3
uWSGI>=2.0.19.1,<2.1
